Move all internal types to `types.ts`
